Fixes for pullrequest
some of the fixes/changes required for the pull request
