+++ pom.xml
@@ -60,6 +60,14 @@
         </dependency>
         <!--end spring security-->
 
+        <!--Mail dependency-->
+        <dependency>
+            <groupId>javax.mail</groupId>
+            <artifactId>mail</artifactId>
+            <version>1.4</version>
+        </dependency>
+        <!--end mail dependency-->
+
         <!-- WebJars -->
         <dependency>
             <groupId>org.webjars</groupId>
@@ -70,11 +78,6 @@
             <groupId>org.webjars</groupId>
             <artifactId>bootstrap</artifactId>
             <version>3.3.1</version>
-        </dependency>
-        <dependency>
-            <groupId>javax.mail</groupId>
-            <artifactId>mail</artifactId>
-            <version>1.4</version>
         </dependency>
         <!-- end webjar -->
     </dependencies>
