+++ src/main/java/com/watchdog/business/User.java
@@ -7,6 +7,7 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.Arrays;
 
+
 /**
  * Created by jmullen on 9/14/16.
  */
