+++ src/main/java/com/watchdog/dao/UserDao.java
@@ -26,5 +26,6 @@ public interface UserDao {
 
     //Get All
     List<User> getAll();
+
 }
