+++ src/main/java/com/watchdog/dao/UserDaoImpl.java
@@ -7,9 +7,7 @@ package com.watchdog.dao;
 import com.watchdog.business.User;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import javax.sql.DataSource;
